refactor(ui): replace Select
