+++ packages/ui/public/tailwind.css
@@ -1136,10 +1136,6 @@ video {
   top: 0.75rem;
 }
 
-.right-2 {
-  right: 0.5rem;
-}
-
 .top-1\/2 {
   top: 50%;
 }
@@ -1158,6 +1154,10 @@ video {
 
 .-right-3 {
   right: -0.75rem;
+}
+
+.right-2 {
+  right: 0.5rem;
 }
 
 .right-4 {
@@ -1637,10 +1637,6 @@ video {
   min-height: 128px;
 }
 
-.min-h-\[4\.5em\] {
-  min-height: 4.5em;
-}
-
 .min-h-\[7em\] {
   min-height: 7em;
 }
@@ -1669,8 +1665,8 @@ video {
   width: 16rem;
 }
 
-.w-32 {
-  width: 8rem;
+.w-36 {
+  width: 9rem;
 }
 
 .w-max {
@@ -1689,10 +1685,6 @@ video {
 
 .w-11 {
   width: 2.75rem;
-}
-
-.w-36 {
-  width: 9rem;
 }
 
 .w-72 {
@@ -1767,6 +1759,10 @@ video {
   width: 6rem;
 }
 
+.w-32 {
+  width: 8rem;
+}
+
 .w-9 {
   width: 2.25rem;
 }
@@ -1797,10 +1793,6 @@ video {
 
 .min-w-\[200px\] {
   min-width: 200px;
-}
-
-.min-w-\[15rem\] {
-  min-width: 15rem;
 }
 
 .max-w-2xs {
@@ -1995,16 +1987,16 @@ video {
   animation: spin 1s linear infinite;
 }
 
+.cursor-not-allowed {
+  cursor: not-allowed;
+}
+
 .cursor-pointer {
   cursor: pointer;
 }
 
 .cursor-default {
   cursor: default;
-}
-
-.cursor-not-allowed {
-  cursor: not-allowed;
 }
 
 .touch-none {
@@ -2085,6 +2077,10 @@ video {
   justify-content: space-between;
 }
 
+.gap-2 {
+  gap: 0.5rem;
+}
+
 .gap-8 {
   gap: 2rem;
 }
@@ -2099,10 +2095,6 @@ video {
 
 .gap-1 {
   gap: 0.25rem;
-}
-
-.gap-2 {
-  gap: 0.5rem;
 }
 
 .gap-6 {
@@ -2277,10 +2269,6 @@ video {
 
 .overflow-x-hidden {
   overflow-x: hidden;
-}
-
-.overscroll-contain {
-  overscroll-behavior: contain;
 }
 
 .truncate {
@@ -2477,6 +2465,10 @@ video {
   border-color: hsl(var(--tw-colors-gray-700) / var(--tw-border-opacity));
 }
 
+.border-transparent {
+  border-color: transparent;
+}
+
 .border-primary-700 {
   --tw-border-opacity: 1;
   border-color: hsl(var(--tw-colors-primary-700) / var(--tw-border-opacity));
@@ -2505,10 +2497,6 @@ video {
 .border-gray-400 {
   --tw-border-opacity: 1;
   border-color: hsl(var(--tw-colors-gray-400) / var(--tw-border-opacity));
-}
-
-.border-transparent {
-  border-color: transparent;
 }
 
 .border-gray-1100 {
@@ -3003,26 +2991,6 @@ video {
   padding-left: 0.875rem;
 }
 
-.pr-8 {
-  padding-right: 2rem;
-}
-
-.pl-7 {
-  padding-left: 1.75rem;
-}
-
-.pr-4 {
-  padding-right: 1rem;
-}
-
-.pl-1\.5 {
-  padding-left: 0.375rem;
-}
-
-.pl-1 {
-  padding-left: 0.25rem;
-}
-
 .pt-1 {
   padding-top: 0.25rem;
 }
@@ -3049,6 +3017,10 @@ video {
 
 .pb-1 {
   padding-bottom: 0.25rem;
+}
+
+.pl-1 {
+  padding-left: 0.25rem;
 }
 
 .pt-8 {
@@ -3234,14 +3206,14 @@ video {
   letter-spacing: -0.025em !important;
 }
 
-.text-gray-800 {
-  --tw-text-opacity: 1;
-  color: hsl(var(--tw-colors-gray-800) / var(--tw-text-opacity));
-}
-
 .text-gray-1100 {
   --tw-text-opacity: 1;
   color: hsl(var(--tw-colors-gray-1100) / var(--tw-text-opacity));
+}
+
+.text-gray-800 {
+  --tw-text-opacity: 1;
+  color: hsl(var(--tw-colors-gray-800) / var(--tw-text-opacity));
 }
 
 .text-gray-900 {
@@ -3357,11 +3329,6 @@ video {
   color: hsl(var(--tw-colors-violet-900) / var(--tw-text-opacity));
 }
 
-.text-gray-300 {
-  --tw-text-opacity: 1;
-  color: hsl(var(--tw-colors-gray-300) / var(--tw-text-opacity));
-}
-
 .text-primary-1200 {
   --tw-text-opacity: 1;
   color: hsl(var(--tw-colors-primary-1200) / var(--tw-text-opacity));
@@ -3440,6 +3407,12 @@ video {
   opacity: 0.75;
 }
 
+.shadow-none {
+  --tw-shadow: 0 0 #0000;
+  --tw-shadow-colored: 0 0 var(--tw-shadow-color);
+  box-shadow: var(--tw-ring-offset-shadow, 0 0 #0000), var(--tw-ring-shadow, 0 0 #0000), var(--tw-shadow);
+}
+
 .shadow {
   --tw-shadow: 0px 4px 8px -2px rgba(16, 24, 40, 0.1), 0px 2px 4px -2px rgba(16, 24, 40, 0.06);
   --tw-shadow-colored: 0px 4px 8px -2px var(--tw-shadow-color), 0px 2px 4px -2px var(--tw-shadow-color);
@@ -3467,12 +3440,6 @@ video {
 .shadow-xl {
   --tw-shadow: 0px 20px 24px -4px rgba(16, 24, 40, 0.08), 0px 8px 8px -4px rgba(16, 24, 40, 0.03);
   --tw-shadow-colored: 0px 20px 24px -4px var(--tw-shadow-color), 0px 8px 8px -4px var(--tw-shadow-color);
-  box-shadow: var(--tw-ring-offset-shadow, 0 0 #0000), var(--tw-ring-shadow, 0 0 #0000), var(--tw-shadow);
-}
-
-.shadow-none {
-  --tw-shadow: 0 0 #0000;
-  --tw-shadow-colored: 0 0 var(--tw-shadow-color);
   box-shadow: var(--tw-ring-offset-shadow, 0 0 #0000), var(--tw-ring-shadow, 0 0 #0000), var(--tw-shadow);
 }
 
@@ -3575,6 +3542,10 @@ video {
   transition-duration: 150ms;
 }
 
+.duration-150 {
+  transition-duration: 150ms;
+}
+
 .duration-100 {
   transition-duration: 100ms;
 }
@@ -3583,12 +3554,12 @@ video {
   transition-duration: 75ms;
 }
 
-.duration-150 {
-  transition-duration: 150ms;
-}
-
 .duration-200 {
   transition-duration: 200ms;
+}
+
+.ease-in-out {
+  transition-timing-function: cubic-bezier(0.4, 0, 0.2, 1);
 }
 
 .ease-out {
@@ -3597,10 +3568,6 @@ video {
 
 .ease-in {
   transition-timing-function: cubic-bezier(0.4, 0, 1, 1);
-}
-
-.ease-in-out {
-  transition-timing-function: cubic-bezier(0.4, 0, 0.2, 1);
 }
 
 .line-clamp-2 {
@@ -3670,6 +3637,11 @@ video {
   border-radius: 0.25rem;
 }
 
+.hover\:border-gray-800:hover {
+  --tw-border-opacity: 1;
+  border-color: hsl(var(--tw-colors-gray-800) / var(--tw-border-opacity));
+}
+
 .hover\:border-primary-900:hover {
   --tw-border-opacity: 1;
   border-color: hsl(var(--tw-colors-primary-900) / var(--tw-border-opacity));
@@ -3688,16 +3660,6 @@ video {
 .hover\:border-red-800:hover {
   --tw-border-opacity: 1;
   border-color: hsl(var(--tw-colors-red-800) / var(--tw-border-opacity));
-}
-
-.hover\:border-gray-800:hover {
-  --tw-border-opacity: 1;
-  border-color: hsl(var(--tw-colors-gray-800) / var(--tw-border-opacity));
-}
-
-.hover\:border-primary-700:hover {
-  --tw-border-opacity: 1;
-  border-color: hsl(var(--tw-colors-primary-700) / var(--tw-border-opacity));
 }
 
 .hover\:border-gray-400:hover {
@@ -4218,6 +4180,11 @@ video {
   background-color: rgb(22 22 22 / 0.3);
 }
 
+.dark .dark\:bg-gray-300 {
+  --tw-bg-opacity: 1;
+  background-color: hsl(var(--tw-colors-gray-300) / var(--tw-bg-opacity));
+}
+
 .dark .dark\:bg-grayd-900\/25 {
   background-color: rgb(112 112 112 / 0.25);
 }
@@ -4235,31 +4202,6 @@ video {
 .dark .dark\:bg-red-500 {
   --tw-bg-opacity: 1;
   background-color: hsl(var(--tw-colors-red-500) / var(--tw-bg-opacity));
-}
-
-.dark .dark\:bg-red-900 {
-  --tw-bg-opacity: 1;
-  background-color: hsl(var(--tw-colors-red-900) / var(--tw-bg-opacity));
-}
-
-.dark .dark\:bg-red-200 {
-  --tw-bg-opacity: 1;
-  background-color: hsl(var(--tw-colors-red-200) / var(--tw-bg-opacity));
-}
-
-.dark .dark\:bg-gray-100 {
-  --tw-bg-opacity: 1;
-  background-color: hsl(var(--tw-colors-gray-100) / var(--tw-bg-opacity));
-}
-
-.dark .dark\:bg-gray-300 {
-  --tw-bg-opacity: 1;
-  background-color: hsl(var(--tw-colors-gray-300) / var(--tw-bg-opacity));
-}
-
-.dark .dark\:bg-gray-400 {
-  --tw-bg-opacity: 1;
-  background-color: hsl(var(--tw-colors-gray-400) / var(--tw-bg-opacity));
 }
 
 .dark .dark\:bg-opacity-70 {
@@ -4359,6 +4301,11 @@ video {
 .dark .dark\:focus\:bg-gray-900:focus {
   --tw-bg-opacity: 1;
   background-color: hsl(var(--tw-colors-gray-900) / var(--tw-bg-opacity));
+}
+
+.dark .dark\:focus\:bg-gray-500:focus {
+  --tw-bg-opacity: 1;
+  background-color: hsl(var(--tw-colors-gray-500) / var(--tw-bg-opacity));
 }
 
 .dark .dark\:focus\:text-gray-100:focus {
