+++ packages/ui/src/components/form-field/form-field.tsx
@@ -3,12 +3,13 @@ import { AnimatePresence, m } from 'framer-motion';
 import React from 'react';
 
 import { easeInOutOpacity } from '../animation';
+import type { InputProps } from '../input';
+import type { SelectProps } from '../select';
 import { Text } from '../text';
-import { InputProps } from './input';
-import { SelectProps } from './select';
+import type { TextAreaProps } from '../text-area';
 
 export type FormFieldProps = {
-  children: React.ReactElement<InputProps | SelectProps>;
+  children: React.ReactElement<InputProps | SelectProps | TextAreaProps>;
   label?: string;
   placeholder?: string;
   errorMessage?: string;
