+++ packages/ui/src/components/form-field/index.ts
@@ -1,4 +1,1 @@
 export * from './form-field';
-export * from './input';
-export * from './select';
-export * from './text-area';
