--- packages/ui/src/components/form-field/styles.ts
@@ -1,7 +1,0 @@
-export const styles = {
-  disabled: 'cursor-not-allowed bg-gray-300 text-gray-1100 shadow-none',
-  borderHover: 'hover:border-gray-800',
-  focus:
-    'focus:outline-none focus-visible:ring-4 focus-visible:ring-primary-700 focus-visible:ring-opacity-25',
-  transition: 'transition duration-150 ease-in-out',
-};
