--- packages/ui/src/components/select/select.module.scss
@@ -1,3 +1,0 @@
-.icon {
-  --geist-icons-background: theme('colors.gray.1100');
-}
